circhelp: Circular Analyses Helper Functions

Light-weight functions for computing descriptive statistics in different circular spaces (e.g., 2pi, 180, or 360 degrees), to handle angle-dependent biases, pad circular data, and more. Specifically aimed for psychologists and neuroscientists analyzing circular data. Basic methods are based on Jammalamadaka and SenGupta (2001) doi:10.1142/4031 , removal of cardinal biases is based on the approach introduced in van Bergen, Ma, Pratte, & Jehee (2015) doi:10.1038/nn.4150 and Chetverikov and Jehee (2023) doi:10.1038/s41467-023-43251-w .

See also

Author

Maintainer: Andrey Chetverikov andrey.chetverikov@uib.no (ORCID)

Other contributors:

Eline Van Geert eline.vangeert@kuleuven.be (ORCID) [contributor]