Remove cardinal biases

Remove cardinal biases

remove_cardinal_biases(
  err,
  x,
  space = "180",
  bias_type = "fit",
  do_plot = FALSE,
  poly_deg = 4,
  var_sigma = TRUE,
  var_sigma_poly_deg = 4,
  reassign_at_boundaries = TRUE,
  reassign_range = 2,
  debug = FALSE
)

Arguments

err

a vector of errors, deviations of response from the true stimuli

x

a vector of true stimuli in degrees (see space)

space

circular space to use (a string: '180' or '360')

bias_type

bias type to use ('fit', 'card', or 'obl', see details)

do_plot

show plots (default: False)

poly_deg

degree of the fitted polynomials for each bin (default: 4)

var_sigma

allow standard deviation (width) of the fitted response distribution to vary as a function of distance to the nearest cardinal (default: True)

var_sigma_poly_deg

degree of the fitted polynomials for each bin for the first approximation for the response distribution to select the best fitting model (default: 4)

reassign_at_boundaries

select the bin for the observations at the boundaries between bins based on the best-fitting polynomial (default: True)

reassign_range

maximum distance to the boundary at which reassignment can occur (default: 2 degrees)

debug

print some extra info (default: False)

Value

a list with the follwing elements:

is_outlier - 0 for outliers (defined as ±3*pred_sigma for the model with varying sigma or as ±3\*SD for the simple model)

pred predicted error

be_c error corrected for biases (be_c = observed error - pred)

which_bin the numeric ID of the bin that the stimulus belong to

bias the bias computed as described above

bias_typ bias type (cardinal or oblique)

pred_lin predicted error for a simple linear model for comparison

pred_sigma predicted SD of the error distribution

coef_sigma_int, coef_sigma_slope intercept and slope for the sigma prediction

Details

If the bias_type is set to 'fit', the function computes the cardinal biases in the following way:

Create two sets of bins, splitting the stimuli vector into bins centered at cardinal and at oblique directions.

For each set of bins, fit a nth-degree polynomial for the responses in each bin, optionally allowing the distribution of responses to vary in width as a function of distance to the nearest cardinal (regardless of whether the bins are centered at the cardinal or at the oblique, the width of the response distribution usually increases as the distance to cardinals increase).

Choose the best-fitting model between the one using cardinal and the one using oblique bins.

Compute the residuals of the best-fitting model - that's your bias-corrected error - and the biases (see below).

The bias is computed by flipping the sign of errors when the average predicted error is negative, so, that, for example, if on average the responses are shifted clockwise relative to the true values, the trial-by-trial error would count as bias when it is also shifted clockwise.

If bias_type is set to 'obl' or 'card', only one set of bins is used, centred at cardinal or oblique angles, respectively.

Examples

# Data in orientation domain from Pascucci et al. (2019, PLOS Bio),
# https://doi.org/10.5281/zenodo.2544946

data <- fread('https://zenodo.org/record/2544946/files/Experiment2_rawdata.csv?download=1')
data[, err:=angle_diff_180(reported, orientation)]
#>       observer orientation reported        rt err
#>    1:        1         135      137 1.0829786   2
#>    2:        1          65       56 0.9887931  -9
#>    3:        1          61       65 1.5067748   4
#>    4:        1          27       25 1.9070205  -2
#>    5:        1          22       20 2.0247443  -2
#>   ---                                            
#> 4396:       10          35       26 1.7775651  -9
#> 4397:       10         141      135 2.0365374  -6
#> 4398:       10         178      163 1.1301296 -15
#> 4399:       10         168      168 1.3772832   0
#> 4400:       10          24       28 2.3897599   4
data[observer==4, remove_cardinal_biases(err, orientation, do_plot = TRUE)]
#> Loading required package: ggplot2
#> Warning: There is a discrepancy  between the original and the re-fit 
#>  used to achieve 'safe' predictions 
#>  
#> Warning: There is a discrepancy  between the original and the re-fit 
#>  used to achieve 'safe' predictions 
#>  
#> Error in data[, circ_sd_fun(err)]: object of type 'closure' is not subsettable

# Data in motion domain from Bae & Luck (2018, Neuroimage), https://osf.io/2h6w9/
data_motion <- fread('https://osf.io/4m2kb/download')
data_motion[, err:=angle_diff_360(RespAngle, TargetDirection)]
#>        subject_Num trial_Num TargetDirection RespAngle motionCoh AngularError
#>     1:         803         1           267.0    255.00     0.512        11.80
#>     2:         803         2            74.2     35.60     0.256        38.60
#>     3:         803         3           336.0    340.00     0.256        -4.33
#>     4:         803         4            70.4     73.80     0.512        -3.34
#>     5:         803         5           109.0    114.00     0.512        -4.42
#>    ---                                                                       
#> 20476:         819       636            14.8      9.92     0.512         4.86
#> 20477:         819       637           130.0    148.00     0.512       -18.10
#> 20478:         819       638           190.0    213.00     0.512       -23.40
#> 20479:         819       639           104.0    118.00     0.256       -14.30
#> 20480:         819       640            15.8     33.60     0.512       -17.90
#>        TarBinID BinCenter Duration Block Session    err
#>     1:       13     270.0      1.5     1       1 -12.00
#>     2:        4      67.5      1.5     1       1 -38.60
#>     3:       16     338.0      1.5     1       1   4.00
#>     4:        4      67.5      1.5     1       1   3.40
#>     5:        6     113.0      1.5     1       1   5.00
#>    ---                                                 
#> 20476:        2      22.5      1.5    10       2  -4.88
#> 20477:        7     135.0      1.5    10       2  18.00
#> 20478:        9     180.0      1.5    10       2  23.00
#> 20479:        6     113.0      1.5    10       2  14.00
#> 20480:        2      22.5      1.5    10       2  17.80
data_motion[subject_Num == unique(subject_Num)[5],
remove_cardinal_biases(err, TargetDirection, space = '360', do_plot = TRUE)]
#> Warning: There is a discrepancy  between the original and the re-fit 
#>  used to achieve 'safe' predictions 
#>  
#> Warning: There is a discrepancy  between the original and the re-fit 
#>  used to achieve 'safe' predictions 
#>  
#> Warning: There is a discrepancy  between the original and the re-fit 
#>  used to achieve 'safe' predictions 
#>  
#> Warning: There is a discrepancy  between the original and the re-fit 
#>  used to achieve 'safe' predictions 
#>  
#> Error in data[, circ_sd_fun(err)]: object of type 'closure' is not subsettable